requirements.txt
alembic==1.7.7
anyio==3.5.0
apache-airflow==2.2.5
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-google==6.7.0
apache-airflow-providers-http==1.0.0
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-postgres==4.1.0
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.1
cachelib==0.6.0
cachetools==4.2.2
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.0
croniter==1.3.4
cryptography==36.0.2
db-dtypes==1.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.4
dnspython==2.2.1
docutils==0.16
email-validator==1.1.3
eventlet==0.33.0
filelock==3.6.0
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
gevent==21.12.0
google==3.0.0
google-ads==14.0.0
google-api-core==1.31.5
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud==0.34.0
google-cloud-aiplatform==1.11.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-automl==2.7.2
google-cloud-bigquery==2.34.2
google-cloud-bigquery-datatransfer==3.6.1
google-cloud-bigquery-storage==2.13.0
google-cloud-bigtable==1.7.0
google-cloud-build==3.8.1
google-cloud-container==1.0.1
google-cloud-core==1.7.2
google-cloud-datacatalog==3.7.1
google-cloud-dataplex==0.2.1
google-cloud-dataproc==4.0.1
google-cloud-dataproc-metastore==1.5.0
google-cloud-dlp==1.0.0
google-cloud-kms==2.11.1
google-cloud-language==1.3.0
google-cloud-logging==3.0.0
google-cloud-memcache==1.3.1
google-cloud-monitoring==2.9.1
google-cloud-orchestration-airflow==1.3.1
google-cloud-os-login==2.6.1
google-cloud-pubsub==2.11.0
google-cloud-redis==2.8.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.44.0
google-cloud-tasks==2.8.1
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.6.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.19.1
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.45.0
grpcio-gcp==0.2.2
grpcio-status==1.45.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.19.1
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==3.0.3
joblib==1.1.0
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
looker-sdk==22.4.0
lxml==4.8.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
multitasking==0.0.10
nltk==3.7
nox==2020.12.31
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
pandas-gbq==0.17.4
pendulum==2.1.2
platformdirs==2.5.1
prison==0.2.1
proto-plus==1.18.1
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.11.2
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.0.1
rsa==4.8
secure-smtplib==0.1.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.24
sqlalchemy-bigquery==1.4.3
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tqdm==4.64.0
typing-extensions==4.1.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.9
virtualenv==20.14.0
Werkzeug==1.0.1
wrapt==1.14.0
WTForms==2.3.3
yfinance==0.1.70
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
